+++ domain/src/main/java/nexters/payout/domain/portfolio/domain/repository/PortfolioStockRepository.java
@@ -1,6 +1,6 @@
 package nexters.payout.domain.portfolio.domain.repository;
 
-import nexters.payout.domain.portfolio.PortfolioStock;
+import nexters.payout.domain.portfolio.domain.PortfolioStock;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
